requirements.txt
Django<1.6
South<2.0
Markdown<2.4
Pygments<2.0
django-model-utils<1.5
